refactor : member service에서 affiliationSerivce로 분리
